Update TimeSeries.php
Should've been floats.

+++ src/TimeSeries.php
@@ -296,8 +296,8 @@ final class TimeSeries
      * @param int|null $count
      * @param AggregationRule|null $rule
      * @param bool $reverse
-     * @param int|null $filter_by_value_from
-     * @param int|null $filter_by_value_to
+     * @param float|null $filter_by_value_from
+     * @param float|null $filter_by_value_to
      * @return Sample[]
      * @throws RedisClientException
      * @throws RedisException
@@ -309,8 +309,8 @@ final class TimeSeries
         ?int $count = null,
         ?AggregationRule $rule = null,
         bool $reverse = false
-        ?int $filter_by_value_from = null,
-        ?int $filter_by_value_to = null
+        ?float $filter_by_value_from = null,
+        ?float $filter_by_value_to = null
     ): array
     {
         $fromTs = $from ? (string)DateTimeUtils::timestampWithMsFromDateTime($from) : '-';
@@ -325,8 +325,8 @@ final class TimeSeries
 
         if ($filter_by_value_from !== null && $filter_by_value_to !== null) {
             $params[] = 'FILTER_BY_VALUE';
-            $params[] = (int)$filter_by_value_from;
-            $params[] = (int)$filter_by_value_to;
+            $params[] = (float)$filter_by_value_from;
+            $params[] = (float)$filter_by_value_to;
         }
 
         $rawResults = $this->redis->executeCommand(array_merge($params, $this->getAggregationParams($rule)));
